### 2025-05-17 #8247 @editorialbot
[PRE REVIEW]: MotilA – A Python pipeline for the analysis of microglial fine process motility in 3D time-lapse multiphoton microscopy data
**Submitting author:** @FabrizioMusacchio (<a href="http://orcid.org/0000-0002-9043-3349">Fabrizio Musacchio</a>)
**Repository:** https://github.com/FabrizioMusacchio/MotilA
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.2
