### 2025-05-17 #8247 @arfon
If you want to fix the DOIs (noting that @editorialbot's suggestions are not always right), you can, then use the following commands (one at a time, as the first line of a new comment) to regenerate the PDF and check the references.
