### 2025-05-19 872d788
Add script reference for figure generation in manuscript and provide dataset details
