+++ paper/paper.bib
@@ -99,7 +99,7 @@
     number = {12},
     pages = {932-940},
     year = {2005},
-    doi = {10.1038/35057065},
+    doi = {10.1038/nmeth818},
     URL = {https://www.nature.com/articles/nmeth818},
     eprint = {https://www.nature.com/articles/nmeth818.pdf},
 }
@@ -353,7 +353,8 @@ doi = {10.1364/OL.33.000156},
   volume    = {154},
   pages     = {165--181},
   year      = {1996},
-  publisher = {Elsevier}
+  publisher = {Elsevier},
+  doi       = {10.1016/0304-3975(94)00262-2}
 }
 
 @techreport{Wu:2005,
@@ -363,7 +364,8 @@ doi = {10.1364/OL.33.000156},
   year        = {2005},
   number      = {LBNL-56864},
   url         = {http://repositories.cdlib.org/lbnl/LBNL-56864},
-  type        = {Technical Report}
+  type        = {Technical Report},
+  doi         = {10.1117/12.596105},
 }
 
 % On the human BIAS in data/image analysis:
@@ -403,7 +405,8 @@ doi = {10.1364/OL.33.000156},
   journal={2016 IEEE Conference on Computer Vision and Pattern Recognition (CVPR)},
   year={2015},
   pages={2930-2939},
-  url={https://api.semanticscholar.org/CorpusID:3039286}
+  url={https://api.semanticscholar.org/CorpusID:3039286},
+  doi={10.1109/cvpr.2016.320},
 }
 
 @article{Brown:2017,
