### 2025-05-19 #8247 @FabrizioMusacchio
@arfon Thanks for letting me know and for the helpful info and instructions! I'll go ahead and fix the missing and incorrect DOIs now and then run the suggested bot commands.
