### 2025-05-19 0f8bef6
Add overview image and description of MotilA pipeline steps in README
