### 2025-05-19 4e1e1ec
Add overview image and description of MotilA pipeline steps in README
