### 2025-05-27 df0847c
Update zarr array creation to support version 3 compatibility
