### 2025-05-27 e22bd90
Enhance TIFF to Zarr conversion with error handling and support for Zarr version 3
