+++ motila/motila.py
@@ -579,50 +579,83 @@ def extract_subvolume(fname, I_shape, projection_layers, projection_range, log,
         raise ValueError("Currently, only TIFF files are supported.")
     else:
         # open the tif file as zarr store:
-        store = tifffile.imread(fname, aszarr=True)
-        I = zarr.open(store, mode='r')
-        
-        # ensure that the zarr store is indeed memory-mapped and not in-memory loaded (which is the default in tifffile):
-        if two_channel:
-            chunks = (1, 1, 1, I_shape[-2], I_shape[-1])
-        else:
-            chunks = (1, 1, I_shape[-2], I_shape[-1])
-        zarr_path = Path(fname).parent.joinpath(Path(fname).stem + ".zarr")
-        #compressor = Blosc(cname='lz4', clevel=5, shuffle=Blosc.SHUFFLE, blocksize=0)
-        #compressor = Blosc(cname='zlib', clevel=9, shuffle=Blosc.SHUFFLE, blocksize=0)
-        #compressor = Blosc(cname='zstd', clevel=9, shuffle=Blosc.SHUFFLE, blocksize=0)
-        # info: we do not compress for speed reasons
-        zarr_group = zarr.group(zarr_path, overwrite=True)
-        if zarr.__version__ >= "3":
-            zarr_group.create_array("image", shape=I.shape, chunks=chunks, 
-                                    dtype=I.dtype)
-        else:    
-            zarr_group.create_dataset("image", shape=I.shape, chunks=chunks, 
-                                    dtype=I.dtype)
-        zarr_group["image"][:] = I
-        zarr_group.attrs["original_file"] = str(fname)
-        zarr_group.attrs["ZARR file path"] = str(zarr_path)
-        zarr_group.attrs["shape"] = I.shape
-        zarr_group.attrs["dtype"] = str(I.dtype)
-        store.close()
+        # tifffile's zarr support works a) for ZARR >= 3.0 and b) not within a Jupyter notebook due to
+        # ZARR asynchronous loading of the data, which is not supported in Jupyter notebooks. Thus, we
+        # use the following try-except block to handle this:
+        try:
+            store = tifffile.imread(fname, aszarr=True)
+            I = zarr.open(store, mode='r')
+            
+            # ensure that the zarr store is indeed memory-mapped and not in-memory loaded (which is the default in tifffile):
+            if two_channel:
+                chunks = (1, 1, 1, I_shape[-2], I_shape[-1])
+            else:
+                chunks = (1, 1, I_shape[-2], I_shape[-1])
+            zarr_path = Path(fname).parent.joinpath(Path(fname).stem + ".zarr")
+            #compressor = Blosc(cname='lz4', clevel=5, shuffle=Blosc.SHUFFLE, blocksize=0)
+            #compressor = Blosc(cname='zlib', clevel=9, shuffle=Blosc.SHUFFLE, blocksize=0)
+            #compressor = Blosc(cname='zstd', clevel=9, shuffle=Blosc.SHUFFLE, blocksize=0)
+            # info: we do not compress for speed reasons
+            zarr_group = zarr.group(zarr_path, overwrite=True)
+            if zarr.__version__ >= "3":
+                zarr_group.create_array("image", shape=I.shape, chunks=chunks, 
+                                        dtype=I.dtype, overwrite=True)
+            else:
+                zarr_group.create_dataset("image", shape=I.shape, chunks=chunks, 
+                                        dtype=I.dtype)
+            zarr_group["image"][:] = I
+            zarr_group.attrs["original_file"] = str(fname)
+            zarr_group.attrs["ZARR file path"] = str(zarr_path)
+            zarr_group.attrs["shape"] = I.shape
+            zarr_group.attrs["dtype"] = str(I.dtype)
[…]
