### 2025-05-27 847dc65
Update Zarr array creation to support version 3 compatibility in subvolume extraction
