### 2025-05-27 f562617
Update spectral unmixing to support Zarr version 3 with conditional dataset/array creation
