### 2025-05-27 fe64f18
Update single_slice_median_filtering to support Zarr version 3 array creation
