### 2025-05-27 05f2026
Update single_slice_circular_median_filtering to support Zarr version 3 array creation
