### 2025-05-27 151c696
Update process_stack to support Zarr version 3 array creation with error handling
