### 2025-06-04 fc0d564
Update dependencies in README and workflow to remove Zarr version restriction
