### 2025-06-04 20f355b
Update spectral_unmix to handle Zarr version 3 array processing
