### 2025-06-04 2e225cf
Update test cases to support Zarr version 3 array creation
