### 2025-06-04 5c6c818
Update tests to support Zarr version 3 array type handling
