### 2025-06-04 22fb6bd
Enhance Zarr compatibility checks in extract_and_register_subvolume test
