### 2025-06-05 77a002b
Add pyproject.toml for project configuration and dependencies
