### 2025-06-05 a27bb7f
Update README.md to enhance installation instructions and add PyPI installation method
