### 2025-06-05 b960b06
Update README.md to enhance installation instructions and add PyPI installation method
