### 2025-06-05 fc54af0
Add BIDS-inspired folder hierarchy description to README.md for improved batch processing and metadata association
