### 2025-07-17 040c2bb
Fix shift application in extract_and_register_subvolume and update dataset creation for Zarr 3.0 compatibility
