### 2025-07-24 910285e
Add pystackreg to project dependencies
