### 2025-07-24 ad7c912
Add pystackreg integration for 2D image registration in reg_2D_images function
