+++ motila/motila.py
@@ -119,7 +119,8 @@ import matplotlib.pyplot as plt
 from matplotlib import colors as mcol
 
 from skimage.registration import phase_cross_correlation
-from skimage import transform
+from skimage import transform, io, exposure
+from pystackreg import StackReg
 import tifffile
 
 import skimage.filters as filter
@@ -1632,7 +1633,8 @@ def plot_intensities(MG_pro, log, plot_path, I_shape):
     return intensity_means
 
 def reg_2D_images(MG_pro, I_shape, log, histogram_ref_stack, max_xy_shift_correction=50,
-                  median_filter_projections=None, median_filter_window_projections=3):
+                  median_filter_projections=None, median_filter_window_projections=3,
+                  usepystackreg=False):
     """
     Registers 2D z-projection images using phase cross-correlation.
 
@@ -1692,26 +1694,46 @@ def reg_2D_images(MG_pro, I_shape, log, histogram_ref_stack, max_xy_shift_correc
     else:
         MG_pro_medianfiltered = MG_pro.copy()
     
+    if usepystackreg:
+        sr = StackReg(StackReg.TRANSLATION)
+        ref = MG_pro_medianfiltered[histogram_ref_stack]
     for stack in range(I_shape[0]):
         # stack=0
-        shifts, _, _ = phase_cross_correlation(reference_image=MG_pro_medianfiltered[histogram_ref_stack],
-                                         moving_image=MG_pro_medianfiltered[stack],
-                                         upsample_factor=30)
-        # check if the shifts are within the allowed range:
-        if shifts[0] > max_xy_shift_correction:
-            shifts[0] = max_xy_shift_correction
-        elif shifts[0] < -max_xy_shift_correction:
-            shifts[0] = -max_xy_shift_correction
-        if shifts[1] > max_xy_shift_correction:
-            shifts[1] = max_xy_shift_correction
-        elif shifts[1] < -max_xy_shift_correction:
-            shifts[1] = -max_xy_shift_correction
-        all_shifts_xy[stack, :] = shifts
-                
-        # apply the shift to the current slice:
-        MG_pro_bin_reg[stack, :, :] = transform.warp(MG_pro_medianfiltered[stack],
-                                                     transform.SimilarityTransform(translation=shifts),
-                                                     preserve_range=True)
+        if not usepystackreg:
+            # use phase cross-correlation to find the shifts:
+            shifts, _, _ = phase_cross_correlation(reference_image=MG_pro_medianfiltered[histogram_ref_stack],
+                                            moving_image=MG_pro_medianfiltered[stack],
+                                            upsample_factor=30)
+            # check if the shifts are within the allowed range:
+            if shifts[0] > max_xy_shift_correction:
+                shifts[0] = max_xy_shift_correction
+            elif shifts[0] < -max_xy_shift_correction:
+                shifts[0] = -max_xy_shift_correction
+            if shifts[1] > max_xy_shift_correction:
+                shifts[1] = max_xy_shift_correction
+            elif shifts[1] < -max_xy_shift_correction:
+                shifts[1] = -max_xy_shift_correction
+            all_shifts_xy[stack, :] = shifts
+                    
+            # apply the shift to the current slice:
+            MG_pro_bin_reg[stack, :, :] = transform.warp(MG_pro_medianfiltered[stack],
+                                                        transform.SimilarityTransform(translation=shifts),
+                                                        preserve_range=True)
+            
+            log.log(f"   phase cross-correlation: registered stack {stack} with shifts {all_shifts_xy[stack, :]}")
+        else:
+            # use pystackreg to find the shifts (eg reg = StackReg(StackReg.TRANSLATION).register_transform(ref,mov)):
[…]
