### 2025-07-24 e3afb25
Enhance reg_2D_images and process_stack functions to support pystackreg for 2D image registration
