### 2025-07-24 1bd3c55
Update README.md to include pystackreg option for 2D registration and clarify registration parameters
