### 2025-07-24 579ef9a
Bump version to 1.0.4 in pyproject.toml
