### 2025-07-24 ec67790
Update installation instructions in README.md to include pystackreg dependency
