### 2025-07-24 5f29136
Add pystackreg to dependencies in Python tests workflow
