### 2025-07-28 0197950
Refactor calc_projection_range to return adjusted projection layers and update documentation
