### 2025-07-28 a408ff1
Update calc_projection_range to return number of layers and adjust tests accordingly
