### 2025-07-28 00f217f
Bump version to 1.0.5 in pyproject.toml
