### 2025-07-28 ce28bfe
Add validation for projection layers in process_stack function
