### 2025-07-28 2864939
Bump version to 1.0.6 in pyproject.toml
