### 2025-07-29 b3cb9df
Fix metadata file filtering in batch_process_stacks function In the batch-processor, we had "metadata.xls" hard coded in _, metadata_file, _ = filterfiles_by_string(Current_TP_Folder, "metadata.xls"). Corrected to: _, metadata_file, _ = filterfiles_by_string(Current_TP_Folder, metadata_file)
