+++ motila/motila.py
@@ -2972,7 +2972,7 @@ def batch_process_stacks(PROJECT_Path, ID_list=[], project_tag="TP000", reg_tif_
                 
             
             # search for metadata file in the current TP_folder and extract the parameters from it (if any):
-            _, metadata_file, _ = filterfiles_by_string(Current_TP_Folder, metadata_file)
+            _, metadata_file, _ = filterfiles_by_string(Current_TP_Folder, "metadata.xls")
             # remove dot-files from metadata_file list:
             metadata_file = [file for file in metadata_file if not file.startswith(".")]
             projection_centers_use = []
