### 2025-07-30 405fed6
Add usepystackreg parameter to batch_process_stacks function
