### 2025-07-30 bc74974
Bump version to 1.0.7 in pyproject.toml
