### 2025-08-04 #8247 @FabrizioMusacchio
Hi @arfon ,

thanks again for your helpful feedback and the update regarding the waitlist. I just wanted to check in briefly and ask whether there’s any rough estimate as to when the review process for our submission might begin.

Also, just to confirm: I’d like to upload the manuscript to a preprint server such as bioRxiv in parallel. As far as I understand [JOSS policy](https://joss.readthedocs.io/en/latest/submitting.html#preprint-policy), this is entirely compatible – but please let me know if there’s anything I should keep in mind when doing so.

Thanks again and best wishes,
Fabrizio
