### 2025-08-18 300faa9
Update citation in README to reflect preprint and DOI information
