### 2025-09-13 #8247 @FabrizioMusacchio
Hi @csoneson , thanks a lot for handling my submission and for your feedback. You  are absolutely right, my paper is currently far too long compared to the JOSS guidelines. I will carefully shorten it and resubmit within the recommended length. I may need a few days for this. I will inform you here as soon as I have the updated version.

Best wishes,
Fabrizio
