### 2025-09-13 3921060
Update paper.md: shortening to under 1000 words; replaced figure with updated shorter figure
