+++ paper/paper.md
@@ -40,243 +40,47 @@ date: 25 March 2025
 bibliography: paper.bib
 ---
 
-# Summary
-*MotilA* is a Python-based image analysis pipeline for quantifying fine process motility of microglia from 3D time-lapse two-channel fluorescence microscopy data. Developed for high-resolution multiphoton *in vivo* imaging datasets, *MotilA* enables both single-file and batch processing across multiple experimental conditions. It performs image preprocessing, segmentation, and motility quantification over time, using a pixel-based change detection strategy that yields biologically interpretable metrics such as the turnover rate (TOR) of microglial fine processes. While originally designed for microglial imaging, the pipeline can be extended to other cell types and imaging applications that require analysis of dynamic morphological changes. *MotilA* is openly available, platform-independent, and includes extensive documentation, tutorials, and example data to facilitate adoption by the broader scientific community. It is released under the GPL-3.0 open-source license.
 
-# Statement of need
-Microglia are innate immune cells of the central nervous system and exhibit highly dynamic, motile processes that continuously scan their environment [@Nimmerjahn:2005; @Fuhrmann:2010; @Tremblay:2010; @Prinz:2019; @Nebeling:2023]. Quantifying microglial motility at the level of fine processes is crucial for studying their function in health and disease, including neurodegeneration, inflammation, and synaptic remodeling. However, despite the biological importance of this analysis, there is currently no dedicated open-source tool tailored for this task.
+## Summary
+*MotilA* is an open-source Python pipeline for quantifying motility of microglial fine processes in 3D time-lapse two-channel fluorescence microscopy. It was developed for high-resolution *in vivo* multiphoton imaging and supports both single-stack and batch analyses. The workflow performs sub-volume extraction, optional registration and spectral unmixing, 2D z-projection, adaptive segmentation, and pixel-wise change detection across time to compute biologically interpretable metrics such as the fine-process turnover rate (TOR). The code is platform independent, documented with tutorials and an example dataset, and released under GPL-3.0.
 
-To date, researchers typically quantify microglial motility manually (see., e.g., @Nebeling:2023) using general-purpose image processing software such as Fiji/ImageJ  [@Schindelin:2012] or ZEISS ZEN [@ZeissZEN:2025]. While these approaches are well established in the field, they are time-consuming, lack reproducibility, and are not well suited for batch processing or cohort-level comparisons. They often focus on individual microglia, whereas *MotilA* enables analysis of the full field of view, allowing for more comprehensive and scalable quantification. Manual workflows are also more susceptible to human bias, limiting their scalability and objectivity [@Lee:2024; @Wall:2018; @Misra:2015; @Brown:2017].
+## Statement of need
+Microglia are innate immune cells of the central nervous system and continuously remodel highly branched processes to survey brain tissue and to respond to pathology [@Nimmerjahn:2005; @Fuhrmann:2010; @Tremblay:2010; @Prinz:2019]. Quantifying this subcellular motility is important for studies of neuroinflammation, neurodegeneration, and synaptic plasticity. Current practice in many labs relies on manual or semi-manual measurements in general-purpose tools such as Fiji/ImageJ or proprietary software [@Schindelin:2012; @ZeissZEN:2025]. These procedures are time consuming, hard to reproduce, and poorly suited for cohort-level comparisons; they often analyze single cells rather than fields of view, and they are sensitive to user bias [@Wall:2018; @Brown:2017]. There is no dedicated, open, and batch-capable solution tailored to this task.
 
[…]
