### 2025-09-13 565972f
Update paper.md: shortening to under 1000 words; replaced figure with updated shorter figure
