+++ paper/paper.md
@@ -42,10 +42,10 @@ bibliography: paper.bib
 
 
 ## Summary
-*MotilA* is an open-source Python pipeline for quantifying motility of microglial fine processes in 3D time-lapse two-channel fluorescence microscopy. It was developed for high-resolution *in vivo* multiphoton imaging and supports both single-stack and batch analyses. The workflow performs sub-volume extraction, optional registration and spectral unmixing, 2D z-projection, adaptive segmentation, and pixel-wise change detection across time to compute biologically interpretable metrics such as the fine-process turnover rate (TOR). The code is platform independent, documented with tutorials and an example dataset, and released under GPL-3.0.
+*MotilA* is an open-source Python pipeline for quantifying motility of microglial fine processes in 3D time-lapse two-channel fluorescence microscopy. It was developed for high-resolution *in vivo* multiphoton imaging and supports both single-stack and batch analyses. The workflow performs sub-volume extraction, optional registration/unmixing, z-projection, segmentation, and pixel-wise change detection to compute motility metrics such as the turnover rate (TOR). The code is platform independent, documented with tutorials and an example dataset, and released under GPL-3.0.
 
 ## Statement of need
-Microglia are innate immune cells of the central nervous system and continuously remodel highly branched processes to survey brain tissue and to respond to pathology [@Nimmerjahn:2005; @Fuhrmann:2010; @Tremblay:2010; @Prinz:2019]. Quantifying this subcellular motility is important for studies of neuroinflammation, neurodegeneration, and synaptic plasticity. Current practice in many labs relies on manual or semi-manual measurements in general-purpose tools such as Fiji/ImageJ or proprietary software [@Schindelin:2012; @ZeissZEN:2025]. These procedures are time consuming, hard to reproduce, and poorly suited for cohort-level comparisons; they often analyze single cells rather than fields of view, and they are sensitive to user bias [@Wall:2018; @Brown:2017]. There is no dedicated, open, and batch-capable solution tailored to this task.
+Microglia are innate immune cells of the central nervous system and continuously remodel highly branched processes to survey brain tissue and to respond to pathology [@Nimmerjahn:2005; @Fuhrmann:2010; @Tremblay:2010; @Prinz:2019]. Quantifying this subcellular motility is important for studies of neuroinflammation, neurodegeneration, and synaptic plasticity. Current practice in many labs relies on manual or semi-manual measurements in general-purpose tools such as Fiji/ImageJ or proprietary software [@Schindelin:2012; @ZeissZEN:2025]. These procedures are time consuming, hard to reproduce, focus on single cells rather than fields of view, and are sensitive to user bias [@Wall:2018; @Brown:2017]. There is no dedicated, open, and batch-capable solution tailored to this task.
 
 *MotilA* fills this gap with an end-to-end, reproducible pipeline for 3D time-lapse two-channel imaging. It standardizes preprocessing, segmentation, and motility quantification and scales from individual stacks to large experimental cohorts. Although optimized for microglia, the approach generalizes to other motile structures that can be reliably segmented over time.
 
@@ -60,9 +60,9 @@ $$TOR=\frac{G+L}{S+G+L},$$
 
 representing the fraction of pixels that changed between consecutive frames. This pixel-based strategy follows earlier microglial motility work [@Fuhrmann:2010; @Nebeling:2023] while providing a fully automated and batchable implementation with parameter logging and diagnostics.
 
[…]
