+++ paper/paper.md
@@ -42,15 +42,15 @@ bibliography: paper.bib
 
 
 ## Summary
-*MotilA* is an open-source Python pipeline for quantifying motility of microglial fine processes in 3D time-lapse two-channel fluorescence microscopy. It was developed for high-resolution *in vivo* multiphoton imaging and supports both single-stack and batch analyses. The workflow performs sub-volume extraction, optional registration/unmixing, z-projection, segmentation, and pixel-wise change detection to compute motility metrics such as the turnover rate (TOR). The code is platform independent, documented with tutorials and an example dataset, and released under GPL-3.0.
+*MotilA* is an open-source Python pipeline for quantifying microglial fine-process motility in 3D time-lapse two-channel fluorescence microscopy. It was developed for high-resolution *in vivo* multiphoton imaging and supports single-stack and batch analyses. The workflow performs sub-volume extraction, optional registration/unmixing, z-projection, segmentation, and pixel-wise change detection to compute the turnover rate (TOR). The code is platform independent, documented with tutorials and example datasets, and released under GPL-3.0.
 
 ## Statement of need
-Microglia are innate immune cells of the central nervous system and continuously remodel highly branched processes to survey brain tissue and to respond to pathology [@Nimmerjahn:2005; @Fuhrmann:2010; @Tremblay:2010; @Prinz:2019]. Quantifying this subcellular motility is important for studies of neuroinflammation, neurodegeneration, and synaptic plasticity. Current practice in many labs relies on manual or semi-manual measurements in general-purpose tools such as Fiji/ImageJ or proprietary software [@Schindelin:2012; @ZeissZEN:2025]. These procedures are time consuming, hard to reproduce, focus on single cells rather than fields of view, and are sensitive to user bias [@Wall:2018; @Brown:2017]. There is no dedicated, open, and batch-capable solution tailored to this task.
+Microglia are immune cells of the central nervous system and continuously remodel processes to survey brain tissue and respond to pathology [@Nimmerjahn:2005; @Fuhrmann:2010; @Tremblay:2010; @Prinz:2019]. Quantifying this subcellular motility is important for studies of neuroinflammation, neurodegeneration, and synaptic plasticity. Current practice in many labs relies on manual or semi-manual measurements in general-purpose tools such as Fiji/ImageJ or proprietary software [@Schindelin:2012; @ZeissZEN:2025]. These procedures are time consuming, hard to reproduce, focus on single cells, and are sensitive to user bias. [@Wall:2018; @Brown:2017]. There is no dedicated, open, and batch-capable solution tailored to this task.
 
 *MotilA* fills this gap with an end-to-end, reproducible pipeline for 3D time-lapse two-channel imaging. It standardizes preprocessing, segmentation, and motility quantification and scales from individual stacks to large experimental cohorts. Although optimized for microglia, the approach generalizes to other motile structures that can be reliably segmented over time.
 
 ## Implementation and core method
-Input is a 5D stack in TZCYX or TZYX order, where T is time, Z is depth, C is channel, and YX are spatial dimensions. For each time point, *MotilA* extracts a user-defined z-sub-volume, optionally performs 3D motion correction and spectral unmixing, and computes a 2D maximum-intensity projection to enable fast and interpretable segmentation. After adaptive thresholding, the binarized projection at time $t_i$, denoted $B(t_i)$, is compared with $B(t_{i+1})$ to derive a temporal change map
[…]
