### 2025-09-13 #8247 @FabrizioMusacchio
Hi @csoneson

thanks again for your feedback. I have now shortened and revised the paper to comply with the JOSS guidelines (current length: 999 words, excluding references). The updated version, including a simplified figure, is available in the repository.

Please let me know if any further adjustments are needed.

Best wishes,
Fabrizio
