### 2025-09-26 #8247 @csoneson
👋🏻 @Yannan-Chen, @DaniBodor, @arka2696, @pr4deepr - would one of you be interested in reviewing this submission for the Journal of Open Source Software (JOSS)?

> MotilA – A Python pipeline for the analysis of microglial fine process motility in 3D time-lapse multiphoton microscopy data
#8247

Our checklist-driven reviews take place on GitHub, and review guidelines can be found [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).

Thanks in advance!
