### 2025-10-15 #8247 @csoneson
👋🏻 @acsenrafilho, @rmarkello, @peerherholz, @sa-nouri, @craddm, @vigji, @Yannan-Chen, @arka2696 - would one of you be interested in reviewing this submission for the Journal of Open Source Software (JOSS)?

> MotilA – A Python pipeline for the analysis of microglial fine process motility in 3D time-lapse multiphoton microscopy data
#8247

Our checklist-driven reviews take place on GitHub, and review guidelines can be found [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).

Thanks in advance!
