### 2025-10-16 #8247 @acsenrafilho
Hi!

Thanks so much for the invitation to review the project! I'd love to help out with it.

Unfortunately, my schedule is extremely tight right now until the end of October. Would it be a problem if I completed the review in November?

If that timeframe works for you, I'd be happy to accept the invitation! Please let me know.
