### 2025-10-16 #9267 @csoneson
Noting that @acsenrafilho will start the review in November (I will ask the editorial bot to send a reminder in early November, just for me to keep track).
