### 2025-11-04 FabrizioMusacchio/MotilA#1 @niksirbi
[JOSS] Correct and simplify instructions for installation from source
Hi 👋 — first of all, thank you for developing and publishing **MotilA**!

I'm taking a look at it as part of the [JOSS review process](https://github.com/openjournals/joss-reviews/issues/9267).

I was able to install the package successfully from PyPI following the README. However, the instructions for installing from source feel a bit unconventional and could be confusing for users. In particular:

1. **Installing mamba inside a conda environment.**  The guide asks users to install _mamba_ in a conda environment and then use it to install dependencies. This is unusual and arguably unnecessary — since 2023, conda ships with **libmamba** as its default dependency solver, so dependency-resolution speed should no longer require installing `mamba` separately (except for older conda versions). If you have some other reason for it, let me know.
2. **The package never gets installed.**  The README instructs users to clone the repo and install dependencies, but the actual `MotilA` package is never installed into the environment — meaning imports may fail or code changes won’t be recognised.

To align with standard practice in the scientific Python ecosystem, I’d recommend updating the “installation from source” section. For example:

````md
### Installation from source

We recommend using a dedicated virtual environment for development and reproducibility. Here we show an example with conda:

```bash
conda create -n motila python=3.12 -y
conda activate motila
```

Clone the GitHub repository and install the package from its root directory:

```bash
git clone https://github.com/FabrizioMusacchio/MotilA.git
cd MotilA
pip install .
```

Alternatively, install directly from GitHub without cloning:

```bash
pip install git+https://github.com/FabrizioMusacchio/MotilA.git
```

**Note:** If you are contributing or making code changes, install in editable mode so updates are immediately reflected:

```bash
pip install -e .
```

````

This workflow ensures that all dependencies specified in `pyproject.toml` are installed automatically, while also making it easier for contributors to get started.
