### 2025-11-06 eeada34
Updated README.md: enhance installation instructions according recommendation by @niksirbi
