+++ README.md
@@ -49,19 +49,33 @@ pip install motila
 ```
 
 ### Installation from source
-If you prefer to install *MotilA* from source, you can clone or download the GitHub repository:
+We recommend using a dedicated virtual environment for development and reproducibility.
+
+Below is an example using conda:
 
 ```bash
-git clone https://github.com/fabriziomusacchio/MotilA.git
-cd MotilA
+conda create -n motila python=3.12 -y
+conda activate motila
 ```
 
-We recommend setting up a dedicated conda environment for development and reproducibility:
+Clone the GitHub repository and install the package from its root directory:
 
 ```bash
-conda create -n motila python=3.12 mamba -y
-conda activate motila
-mamba install -y numpy scipy matplotlib scikit-image scikit-learn pandas tifffile zarr numcodecs pystackreg openpyxl xlrd ipywidgets ipykernel ipympl
+git clone https://github.com/FabrizioMusacchio/MotilA.git
+cd MotilA
+pip install .
+```
+
+Alternatively, install directly from GitHub without cloning:
+
+```bash
+pip install git+https://github.com/FabrizioMusacchio/MotilA.git
+```
+
+**Note:** If you are contributing or making code changes, install in editable mode so updates are immediately reflected:
+
+```bash
+pip install -e .
 ```
 
 ⚠️ **Avoid mixing install methods**:  
@@ -116,7 +130,7 @@ import motila as mt
 from pathlib import Path
 
 tif_file_path = Path("path/to/your/image_stack.tif")
-corrected_tif_file_path = mt.tiff_axes_check_and_correct(tif_file)
+corrected_tif_file_path = mt.tiff_axes_check_and_correct(tif_file_path)
 ```
 
 The output `corrected_tif_file` is the path to the corrected TIFF file, which is automatically saved in the same directory as the original file.
@@ -252,7 +266,7 @@ You can add several projection centers (`Projection Center 1`, `Projection Cente
 | `projection_layers_default` | integer   | define the number of z-layers to project for motility analysis. |
 | `projection_center_default` | integer   | define the center slice of the projection |
 
-In case of image volumes densely packed with microglia, we recommend to subdivided the volume into several subvolumes with different projection centers. This will help to avoid overlapping microglia in the projection and thus ensure a more accurate capturing of the microglial processes' motility.
+In case of image volumes densely packed with microglia, we recommend to subdivide the volume into several subvolumes with different projection centers. This will help to avoid overlapping microglia in the projection and thus ensure a more accurate capturing of the microglial processes' motility.
 
 Avoid including blood vessels in the projection center. Blood vessels can lead to false-positive motility results, as the pipeline cannot distinguish between microglial processes and blood vessels. 
 
@@ -275,7 +289,7 @@ Avoid including blood vessels in the projection center. Blood vessels can lead t
 | `hist_match`            | bool (`True` or `False`)  | match histograms across 3D stacks |
 | `histogram_ref_stack`   | integer     | define the reference 3D stack for histogram matching. |
 
-Histogram equalization enhances the contrast of the image by stretching the intensity range. This can be particularly useful for images with low contrast or uneven illumination. The `hist_equalization_clip_limit` parameter controls the intensity clipping limit for the histogram equalization. A higher value increases the intensity range but may also amplify noise. The `hist_equalization_kernel_size` parameter defines the kernel size for the histogram equalization. The default is `None` which let's the function choose the kernel size automatically. In cases of occurring block artifacts, you can set a fixed kernel size (e.g., (8,8), (16,16), (24,24), ...).
[…]
