### 2025-11-21 2464485
docs: add CONTRIBUTING.md to guide new contributors on project involvement
