### 2025-11-21 d5ba17c
refactor: streamline module documentation and enhance parameter descriptions
