+++ motila/motila.py
@@ -1,100 +1,20 @@
-""" MotilA: Microglial Motility Analysis Pipeline
-=============================================
+"""
+Core routines for the MotilA pipeline.
 
-MotilA is a Python-based pipeline for analyzing microglial fine process motility from in vivo 
-4D and 5D time-lapse image stacks. It supports both single-file and batch processing modes, 
-and includes optional steps for image registration, spectral unmixing, contrast enhancement, 
-and multi-method segmentation.
+This module implements the main processing functions for microglial fine process
+motility analysis from 4D and 5D time-lapse multiphoton imaging stacks. It 
+provides utilities for loading TIFF/Zarr data, optional subvolume extraction, 
+2D and 3D motion correction, spectral unmixing, intensity normalization, 
+contrast adjustments, filtering, segmentation, and computation of motility 
+metrics such as gain, loss, stability, and turnover rate.
 
-Core pipeline steps:
---------------------
-1. Load image data (supports TIFF in TZCYX and TZYX formats)
-2. Extract sub-volumes from the 3D image stack
-3. (Optional) Register sub-volumes to correct motion
-4. (Optional) Perform spectral unmixing to reduce channel bleed-through
-5. z-projection of the sub-volumes for further processing
-6. (Optional) Apply histogram equalization and histogram matching for contrast/bleaching correction
-7. (Optional) Apply filtering (median, Gaussian) to reduce noise
-8. Segment microglial structures via thresholding and blob detection
-9. Analyze motility by tracking pixel changes over time
+The module exposes three high-level entry points:
 
-Batch processing steps:
------------------------
-1. Define a project folder containing multiple registered TIFF stacks
-2. Apply the core pipeline steps to each stack automatically
-3. Store results (plots, metrics, segmentation) per dataset
-4. Aggregate results across animals or conditions for cohort-level analysis
+* ``process_stack`` for single-stack analysis
+* ``batch_process_stacks`` for automated project-folder processing
+* ``batch_collect`` for aggregation of metrics across datasets
 
-Key features:
--------------
-- Handles 4D and 5D TIFF stacks (TZXY, TCZXY)
-- Supports dual-channel imaging (e.g., microglia and neurons)
-- Implements spectral unmixing, histogram equalization and matching
-- Includes 2D and 3D motion correction
-- Offers multiple thresholding and filtering options
-- Computes gain, loss, stability, and turnover rate (TOR) metrics
-
-Dependencies:
--------------
-- Python >= 3.9
-- numpy, scipy, matplotlib, scikit-image, scikit-learn, pandas
-- tifffile, zarr, numcodecs, openpyxl, xlrd
-- ipywidgets, ipykernel, ipympl
-
-Installation:
--------------
-To install dependencies via conda and mamba:
-```bash
-conda create -n motila python=3.12 mamba -y
-conda activate motila
-mamba install -y numpy scipy matplotlib scikit-image scikit-learn pandas tifffile zarr numcodecs openpyxl xlrd ipywidgets ipykernel ipympl
-```
-
-Usage:
--------------
-1. Configure pipeline parameters (e.g., projection layers, channels, filters).
-2. Run process_stack(fname, ...) for single-file or batch_process_stacks(...) for cohort mode with the desired settings.
-3. Output includes processed stacks, Excel tables with metrics, and diagnostic plots
-
-Author
--------------
-Fabrizio Musacchio (lead developer)
-Contributors: Sophie Crux, Felix Nebeling, Nala Gockel, Falko Fuhrmann, Martin Fuhrmann
-First created: September 2023
-
-License: GPL-3.0 License
--------------
-GPL-3.0 License — see LICENSE file for full terms.
-
-Summary:
-* Free to use, modify, and distribute under the same license
-* Must retain license and attribution
-* Not for use in proprietary/closed-source projects
[…]
