### 2025-11-21 0462604
docs: update parameter and return descriptions in utility functions for clarity
