### 2025-11-21 38ad0a3
1st version of dedicated Docs pages added
