+++ docs/source/overview.rst
@@ -1,0 +1,268 @@
+Overview
+========
+
+MotilA is a Python-based image analysis pipeline designed to quantify microglial
+fine process motility from 4D and 5D time-lapse stacks acquired with in vivo
+multiphoton imaging. While developed for microglia, the pipeline can in principle
+be applied to other cell types and imaging paradigms that share similar data
+characteristics.
+
+MotilA supports both single-file and batch processing, making it suitable for
+small experiments and high-throughput analyses. The main entry points are
+
+* :func:`motila.motila.process_stack` for single-stack analysis,
+* :func:`motila.motila.batch_process_stacks` for batch processing of project
+  folders, and
+* :func:`motila.motila.batch_collect` for aggregation of results across datasets.
+
+
+What does MotilA do?
+--------------------
+
+MotilA automates the processing and analysis of fluorescence microscopy data
+with a focus on microglial process dynamics. The pipeline performs:
+
+* **Preprocessing**: image registration, spectral unmixing, histogram
+  equalization, bleach correction and z-projection of subvolumes.
+* **Segmentation**: adaptive thresholding and noise filtering to isolate
+  microglial fine processes.
+* **Motility quantification**: frame-to-frame analysis of pixel changes in
+  segmented microglial structures.
+* **Batch processing**: automated handling of multiple datasets with
+  standardized parameter settings.
+
+The individual processing steps and parameter choices are documented in more
+detail in the sections on data prerequisites, pipeline steps and parameter
+overview.
+
+
+How is motility determined?
+---------------------------
+
+MotilA quantifies motility by extracting, for each time point :math:`t_i`, a
+3D subvolume around a defined projection center and performing a maximum
+intensity z-projection. This sacrifices z-axis information but allows efficient
+segmentation and frame-to-frame comparison in 2D, which is particularly useful
+for batch processing. This approach is consistent with earlier work on
+microglial motility (for example Nebeling et al. 2023; Fuhrmann et al. 2010).
+
+Let :math:`B(t_i)` denote the binarized projection at time :math:`t_i`. MotilA
+computes a temporal difference
+
+.. math::
+
+   \Delta B(t_i) = 2 \times B(t_i) - B(t_{i+1}) .
+
+Based on the value of :math:`\Delta B(t_i)` at each pixel, three classes are
+defined:
+
+* **Stable pixels** :math:`S` with :math:`\Delta B = 1`,
+* **Gained pixels** :math:`G` with :math:`\Delta B = -1`,
+* **Lost pixels** :math:`L` with :math:`\Delta B = 2`.
+
+From these counts MotilA derives the turnover rate (TOR) as a central motility
+metric,
+
+.. math::
+
+   TOR = \frac{G + L}{S + G + L} ,
+
+which represents the fraction of pixels that change between two consecutive
+time points. Higher TOR values correspond to stronger fine process motility.
+
+
+Pipeline overview
+-----------------
+
+.. figure:: _static/figures/motila_figure_2.png
+   :alt: MotilA pipeline overview
+   :align: center
+   :figwidth: 100%
+
+   **Step-by-step illustration of the MotilA pipeline using the included test 
+   dataset**. **a)** Overview of the image processing pipeline, showing core and 
+   optional steps. **b)** Example projections of a cropped microglial cell at time 
+   points $t_0$ and $t_1$, including raw, histogram-equalized, filtered 
+   (median and Gaussian), and binarized versions.  **c)** Binarized pixel-wise 
+   comparison between $t_0$ and $t_1$, with classification into stable (S, blue), 
+   gained (G, green), lost (L, red), and background (BG, white) pixels, along with 
+   the corresponding pixel statistics.  **d)** Normalized cell brightness over time, 
[…]
