+++ docs/source/tutorials.rst
@@ -1,0 +1,238 @@
+Tutorials and Example Datasets
+==============================
+
+MotilA provides Jupyter notebooks and Python scripts that
+demonstrate the complete analysis workflow. Also, an example dataset is
+publicly available on Zenodo. These resources enable users to
+validate their installation, understand the processing steps and reproduce the
+results shown in the associated manuscript.
+
+Example dataset (Zenodo)
+------------------------
+
+A curated example dataset is available on Zenodo and contains:
+
+* a representative 4D or 5D time-lapse image stack,
+* a corresponding ``metadata.xls`` file with projection center settings, and
+* a ready-to-use project folder structure compatible with MotilA.
+
+The dataset can be downloaded from:
+
+* Gockel & Nieves-Rivera (2025), doi: 10.5281/zenodo.15061566  
+  `Zenodo record <https://zenodo.org/records/15061566>`_
+
+After downloading, place the dataset into the ``example project`` directory of
+MotilA:
+
+* `example project folder <https://github.com/FabrizioMusacchio/MotilA/tree/main/example%20project>`_
+
+Running MotilA directly on this dataset allows users to verify that the pipeline
+is functioning correctly and produces meaningful motility metrics.
+
+Tutorial notebooks
+------------------
+
+MotilA includes Jupyter notebooks that illustrate the core processing steps on
+the example data:
+
+* `single_file_run.ipynb <https://github.com/FabrizioMusacchio/MotilA/blob/main/example%20notebooks/single_file_run.ipynb>`_  
+  Demonstrates the complete workflow for processing a single image stack.
+
+* `batch_run.ipynb <https://github.com/FabrizioMusacchio/MotilA/blob/main/example%20notebooks/batch_run.ipynb>`_  
+  Shows how to process multiple datasets stored in a structured project folder.
+
+These notebooks guide users through image loading, optional preprocessing,
+motility computation and inspection of intermediate and final results. They
+serve as the most accessible introduction to the pipeline.
+
+Example Python scripts
+----------------------
+
+Equivalent Python scripts are provided for users who prefer script-based
+workflows or who want to integrate MotilA into automated analysis pipelines:
+
+* `single_file_run.py <https://github.com/FabrizioMusacchio/MotilA/blob/main/example%20scripts/single_file_run.py>`_
+* `batch_run.py <https://github.com/FabrizioMusacchio/MotilA/blob/main/example%20scripts/batch_run.py>`_
+
+These scripts mirror the behavior of the tutorial notebooks and can be adapted
+for larger projects or command-line environments.
+
+Reproducing manuscript figures
+------------------------------
+
+The figures shown in the associated manuscript were generated using a dedicated
+script that applies MotilA to a specific subset of the example dataset:
+
+* `single_file_run_paper.py <https://github.com/FabrizioMusacchio/MotilA/blob/main/example%20scripts/single_file_run_paper.py>`_
+
+The subset used for figure generation is located at:
+
+* ``example project/Data/ID240103_P17_1_cutout/TP000``
+
+The script contains all parameter settings used during analysis and can be used
+to reproduce the manuscript figures exactly.
+
+Additional example datasets
+---------------------------
+
+The repository may include additional reduced datasets, project templates or
+folder structures intended to help users set up their own analyses. These
+resources are located in the ``example project`` directory and follow the same
+format expected by the batch processing routines of MotilA. They can be used as
+templates for structuring new experimental datasets.
+
+
+
+Example usage
+-------------
+
+The following code examples illustrate how to use MotilA in practice.
+
+Single-file processing
+~~~~~~~~~~~~~~~~~~~~~~
+
+Import MotilA and initialize logging:
+
[…]
