+++ docs/source/index.rst
@@ -1,0 +1,49 @@
+.. MotilA documentation master file, created by
+   sphinx-quickstart on Fri Nov 21 14:13:59 2025.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+MotilA documentation
+====================
+
+
+.. image:: https://img.shields.io/github/v/release/FabrizioMusacchio/motila
+   :target: https://github.com/FabrizioMusacchio/motila/releases
+   :alt: GitHub Release
+
+.. image:: https://img.shields.io/badge/License-GPL%20v3-yellow.svg
+   :target: https://opensource.org/licenses/GPL-3.0
+   :alt: GPLv3 License
+
+.. image:: https://github.com/FabrizioMusacchio/motila/actions/workflows/python-tests.yml/badge.svg
+   :target: https://github.com/FabrizioMusacchio/motila/actions/workflows/python-tests.yml
+   :alt: Test Status
+
+.. image:: https://img.shields.io/pypi/v/motila.svg
+   :target: https://pypi.org/project/motila/
+   :alt: PyPI version
+
+`MotilA <https://github.com/FabrizioMusacchio/MotilA>`_ is a Python-based pipeline for microglial fine process motility analysis
+from 4D and 5D multiphoton imaging data.
+
+
+This documentation provides:
+
+* an overview of the pipeline and its scientific context,
+* installation and basic usage instructions,
+* a detailed description of the processing steps and parameters,
+* an automatically generated API reference.
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents:
+
+   overview
+   data_requirements
+   tutorials
+   parameters
+   api
+   contributing
+
+MotilA is `free and open-source software (FOSS) <https://en.wikipedia.org/wiki/Free_and_open-source_software>`_ distributed under the :ref:`GPL-3.0 license <license>`.
+

[…]
