### 2025-11-21 e10599b
Add Read the Docs configuration file for documentation build
