### 2025-11-21 5a57995
Update Read the Docs configuration to specify Python version 3.12
