+++ .readthedocs.yaml
@@ -6,10 +6,10 @@
 version: 2
 
 # Set the OS, Python version, and other tools you might need
-# build:
-#   os: ubuntu-24.04
-#   tools:
-#     python: "3.13"
+build:
+  os: ubuntu-24.04
+  tools:
+    python: "3.12"
 
 # Build documentation in the "docs/" directory with Sphinx
 sphinx:
