### 2025-11-21 46f8bf0
Shortened README.md after adding an elaborated Read The Docs documentation website
