+++ README.md
@@ -1,10 +1,11 @@
-![GitHub Release](https://img.shields.io/github/v/release/FabrizioMusacchio/motila) [![GPLv3 License](https://img.shields.io/badge/License-GPL%20v3-yellow.svg)](https://opensource.org/licenses/) ![Tests](https://github.com/FabrizioMusacchio/motila/actions/workflows/python-tests.yml/badge.svg) [![PyPI version](https://img.shields.io/pypi/v/motila.svg)](https://pypi.org/project/motila/)
+![GitHub Release](https://img.shields.io/github/v/release/FabrizioMusacchio/motila) [![GPLv3 License](https://img.shields.io/badge/License-GPL%20v3-yellow.svg)](https://opensource.org/licenses/) ![Tests](https://github.com/FabrizioMusacchio/motila/actions/workflows/python-tests.yml/badge.svg) [![PyPI version](https://img.shields.io/pypi/v/motila.svg)](https://pypi.org/project/motila/) [![Documentation Status](https://readthedocs.org/projects/motila/badge/?version=latest)](https://motila.readthedocs.io/en/latest/?badge=latest)
 
 
 
 # MotilA: A pipeline for microglial fine process motility analysis
 
 *MotilA* is a Python-based image analysis pipeline designed to quantify microglial fine process motility from 4D and 5D time-lapse image stacks acquired through multi-photon in vivo imaging. While developed for microglial analysis, *MotilA* can be applied to other cell types and imaging studies as well. The pipeline supports both single-file and batch processing, making it adaptable for various experimental designs and high-throughput analyses. 
+
 
 ## What does MotilA do?
 *MotilA* automates the processing and analysis of fluorescence microscopy data, particularly for microglial process dynamics. It performs:
@@ -13,6 +14,9 @@
 - **Segmentation**: Adaptive thresholding and noise filtering to isolate microglial processes.
 - **Motility quantification**: Frame-to-frame analysis of pixel changes in microglial structures.
 - **Batch processing**: Automated handling of multiple datasets with standardized parameter settings.
+
+A high-level description of all processing steps and parameters is given in the [online documentation](https://motila.readthedocs.io/).
+
 
 ## How is "motility" determined?
 *MotilA* quantifies motility by first extracting a sub-volume from the 3D stack at each imaging time point $t_i$ and performing a maximum intensity z-projection. This sacrifices the z-axis information but enables segmentation and quantification of stable, lost, and gained pixels in a computationally efficient manner, facilitating batch processing with standard image analysis techniques. This approach aligns with methodologies used in prior studies, such as [Nebeling et al. (2023)](https://pubmed.ncbi.nlm.nih.gov/36749020/) or [Fuhrmann et al. (2010)](https://pubmed.ncbi.nlm.nih.gov/20305648/). The temporal variation $\Delta B(t_i)$ is then computed as:
@@ -109,280 +113,8 @@ We used the following Python script to generate the figures presented in our sub
 This script includes all parameter settings used during analysis and can be employed to reproduce the figures. It was applied to a subset of the example dataset described above. This specific subset is available in the repository under [`example project/Data/ID240103_P17_1_cutout/`](https://github.com/FabrizioMusacchio/MotilA/tree/main/example%20project/Data/ID240103_P17_1_cutout/TP000).
 
 
-## Data prerequisites  
-Before using *MotilA*, ensure that your imaging data meets the following requirements:  
 
-### 1. TIFF file format and image axis order  
-*MotilA* expects input image stacks in TIFF format with axes structured as either **TZCYX** (for multi-channel data) or **TZYX** (for single-channel data). These axes correspond to:  
-
[…]
