### 2025-11-21 41d6eae
Clarify license terms in overview.rst to specify that users can modify and adapt the source code to their needs
