### 2025-11-21 7479759
Update license link in README.md to point to the RTD section Remove License section
