### 2025-11-21 9ed8ae2
Update README.md links to point to the correct documentation sections
