### 2025-11-21 1a27a13
Add guidance on defining parameters with links to example notebooks and scripts
