### 2025-11-21 5a5e923
Update README.md to include links to example notebooks and scripts for parameter definition
