### 2025-11-21 117674d
Add guidance on defining parameters with links to example notebooks and scripts
