### 2025-11-22 b8451f2
Add example dataset figure and enhance documentation in tutorials
