+++ docs/source/tutorials.rst
@@ -7,24 +7,31 @@ publicly available on Zenodo. These resources enable users to
 validate their installation, understand the processing steps and reproduce the
 results shown in the associated manuscript.
 
-Example dataset (Zenodo)
+Example dataset
 ------------------------
 
-A curated example dataset is available on Zenodo and contains:
+A curated example dataset is available on Zenodo: Gockel & Nieves-Rivera (2025), `doi: 10.5281/zenodo.15061566 <https://zenodo.org/records/15061566>`_
 
-* a representative 4D or 5D time-lapse image stack,
+The dataset contains:
+
+* a representative 5D time-lapse image stack,
 * a corresponding ``metadata.xls`` file with projection center settings, and
 * a ready-to-use project folder structure compatible with MotilA.
 
-The dataset can be downloaded from:
+.. figure:: _static/figures/zenodo_example_data_content.png
+   :alt: Contents of the Zenodo example dataset and expected project folder structure
+   :align: center
+   :figwidth: 100%
 
-* Gockel & Nieves-Rivera (2025), doi: 10.5281/zenodo.15061566  
-  `Zenodo record <https://zenodo.org/records/15061566>`_
+   Contents of the `Zenodo example dataset <https://zenodo.org/records/15061566>`_, including the required project-folder
+   structure for batch processing. The dataset mirrors the directory layout
+   expected by MotilA, with ``ID`` folders, ``project_tag`` subfolders,
+   registered TIFF stacks, and an accompanying ``metadata.xls`` file.
 
 After downloading, place the dataset into the ``example project`` directory of
-MotilA:
+your MotilA working directory (i.e., the folder, where you run the scripts from). 
 
-* `example project folder <https://github.com/FabrizioMusacchio/MotilA/tree/main/example%20project>`_
+An example of the expected folder structure can also be found `in the repository <https://github.com/FabrizioMusacchio/MotilA/tree/main/example%20project>`_.
 
 Running MotilA directly on this dataset allows users to verify that the pipeline
 is functioning correctly and produces meaningful motility metrics.
@@ -86,7 +93,15 @@ templates for structuring new experimental datasets.
 Example usage
 -------------
 
-The following code examples illustrate how to use MotilA in practice.
+The following examples illustrate how to use MotilA in practice, starting from
+a registered time-lapse stack and proceeding through single-file processing,
+batch processing and batch-level result collection.
+
+MotilA expects input data as TIFF image stacks in either TZCYX or TZYX format,
+with T denoting time, Z depth, C channels and X/Y the spatial dimensions. See
+the :doc:`data requirements <data_requirements>` for details on axis order,
+multi-channel handling and the optional axis correction function
+``tiff_axes_check_and_correct``.
 
 Single-file processing
 ~~~~~~~~~~~~~~~~~~~~~~
@@ -106,8 +121,168 @@ Import MotilA and initialize logging:
    # initialize logger
    log = mt.logger_object()
 
-Then, define the corresponding parameters. A set of example values can be 
-found in the `tutorial notebooks <https://github.com/FabrizioMusacchio/MotilA/tree/main/example%20notebooks>`_ and `scripts <https://github.com/FabrizioMusacchio/MotilA/tree/main/example%20scripts>`_ provided in the repository. 
+Parameter selection
+^^^^^^^^^^^^^^^^^^^
+
+Before calling :func:`motila.process_stack`, you should choose parameter values
+that match your imaging protocol and data quality. A complete description of
+all arguments is given in the :doc:`parameter documentation <parameters>`.
[…]
