### 2025-11-22 6b2c1c1
Add optional development dependencies for testing
