### 2025-11-22 430f0f8
Add pytest configuration for coverage reporting
