+++ pyproject.toml
@@ -42,3 +42,7 @@ dev = [
     "pytest",
     "pytest-cov",
 ]
+
+[tool.pytest.ini_options]
+addopts = "--cov=motila --cov-report=xml"
+testpaths = ["tests"]
