### 2025-11-22 2cf323c
Add pytest-cov for coverage reporting in test workflow
