### 2025-11-22 d42c761
Add coverage reporting and upload to Codecov in test workflow
