### 2025-11-22 da0dad8
Added Codecov badge in README.md
