### 2025-11-22 cf004d3
Added Codecov badge in index.rst
