### 2025-11-22 b551a0a
docs: comment out all sys.path.append lines in tutorials and example scripts
Users installing MotilA via `pip install motila` do not need `sys.path.append(...)`.
These lines are now commented out and marked as optional for local development
from a cloned repository. Updated all tutorial code blocks and example scripts
accordingly to avoid confusion and ensure clean import behavior.
