### 2025-11-22 6e48758
refactor: remove debug __main__ block from motila core module
Removed the legacy debugging/testing code guarded by
`if __name__ == '__main__':` in `motila.py`. Full examples for
single-file and batch processing now live in the dedicated example
scripts and notebooks, keeping the core module focused on the public
API functions.
