+++ motila/motila.py
@@ -3200,223 +3200,8 @@ def batch_collect(PROJECT_Path, ID_list=[], project_tag="TP000", motility_folder
 
 # %% DEBUGGING/TESTING
 if __name__ == '__main__':
-
-    # EXAMPLE PARAMETERS FOR SINGLE FILE RUN:
-
-    # define data input paths:
-    Current_ID = "ID14201"
-    group = "blinded"
-    DATA_Path = "/Users/husker/Science/Python/Projekte/MotilA/example project/Data/ID240103_P17_1/TP000/registered/"
-    IMG_File = "all stacks 4D reg.tif"
-    fname = Path(DATA_Path).joinpath(IMG_File)
-        
-    # define projection settings:
-    projection_layers_default = 44 # define how many z-layers you want to project for the motility analysis
-    projection_center_default = 23 # define the center slice of the projection; a sub-stack of +/- projection_layers will be projected;
-                                   # if metadata.xls is present, this value is ignored (in batch processing only)
-    
-    # define results output path:
-    RESULTS_foldername = f"../motility_analysis_projcenter_{projection_center_default}/"
-    RESULTS_Path = Path(DATA_Path).joinpath(RESULTS_foldername)
-    check_folder_exist_create(RESULTS_Path)
-    clear_previous_results = True # set to True if all files in RESULTS_Path folder should be deleted before processing
-    
-
-    # thresholding settings:
-    threshold_method = "otsu"     # choose from: otsu, li, isodata, mean, triangle, yen, minimum
-    blob_pixel_threshold = 100    # define the threshold for the minimal pixel area of a blob during the segmentation
-    compare_all_threshold_methods = True # if True, all threshold methods will be compared and saved in the plot folder
-
-    # image enhancement settings:
-    hist_equalization = True             # enhance the histograms WITHIN EACH projected stack: True or False
-    hist_equalization_clip_limit = 0.05  # clip limit for the histogram equalization (default is 0.05)
-    hist_equalization_kernel_size = None # kernel size for the histogram equalization; 
-                                         # None (default) for automatic, or use a tuple (x,y) for a fixed size;
-                                         # when using a tuple, you can start increasing the values from multiples
-                                         # of 8, e.g., (8,8), (16,16), (24,24), (32,32), ... (128,128), ...
-    hist_match = True               # match the histograms   ACROSS the stacks          : True or False
-    histogram_ref_stack = 0         # define the stack which should be used as reference for the histogram matching
-    
-    
-    # filter settings:
-    median_filter_slices             = 'circular' # median filter on SLICES BEFORE projecting
-                                            # 'square', 'circular', or False
-                                            # circular: floating point numbers allowed, not lower than 0.5 for circular
-                                            # square: integer numbers (3, 5, 9)
-    median_filter_window_slices      = 1.55 # median filter window size on SLICES BEFORE projecting
-                                            # circular: only values >= 0.5 allowed/have an effect
-                                            # square: integer numbers (3, 5, 9)
-    
-    median_filter_projections        = 'circular' # median filter on PROJECTIONS
-                                            # square, circular, or False
-    median_filter_window_projections = 1.55 # median filter window size on PROJECTIONS
-                                            # circular: only values >= 0.5 allowed/have an effect
-                                            # square: integer numbers (3, 5, 9)
-    gaussian_sigma_proj = 1.00  # Standard deviation of Gaussian (blur) filter applied on the projected stack
-                                # set to 0 for turning off
-
-    # channel settings: 
[…]
