### 2025-11-24 1ab4bb6
Remove unnecessary sys.path modification in example usage of axis correction function
