### 2025-12-01 FabrizioMusacchio/MotilA#1 @niksirbi
@FabrizioMusacchio Along the same lines, I also recommend updating your `python-tests.yml` CI workflow to reflect the new installation method (i.e. using `pip install`). If the CI tests installation differently than what you expect users to do, it won’t properly reflect real-world usage. Speaking from experience, this can lead to situations where users encounter issues that the CI does not catch—and vice versa.
