### 2025-12-01 f59cf8d
Refactor import statements and update installation instructions across documentation and example scripts
- Changed import statements to use `from motila import motila as mt` for consistency.
- Removed unnecessary `sys.path.append` comments from notebooks and scripts.
- Updated installation instructions in README, notebooks, and scripts to direct users to the GitHub repository and documentation for installation guidance.
