### 2025-12-01 e71e6e8
Refactor GitHub Actions workflow for Python tests: update Conda setup and streamline dependency installation (now using pip install .)
