### 2025-12-01 f1c8edd
Refactor package init to expose top-level API and preserve backward compatibility
- Added explicit re-exports of main pipeline functions (process_stack,
  batch_process_stacks, batch_collect, hello_world) in motila/__init__.py
  to enable clean imports via `import motila as mt`.
- Added `from . import motila as motila` to ensure that the full submodule
  `motila.motila` remains available, preserving backward compatibility for
  existing user scripts relying on `from motila import motila as mt`.
- This change keeps all helper functions accessible through the submodule
  import while providing a
[…]
