+++ motila/__init__.py
@@ -1,0 +1,16 @@
+from .motila import (
+    hello_world,
+    process_stack,
+    batch_process_stacks,
+    batch_collect,
+)
+
+__all__ = [
+    "hello_world",
+    "process_stack",
+    "batch_process_stacks",
+    "batch_collect",
+]
+
+# optional: expose the motila module directly
+from . import motila as motila
