+++ example scripts/batch_run.py
@@ -43,7 +43,7 @@ Fabrizio Musacchio,  March 20, 2025
 """
 # %% IMPORTS
 from pathlib import Path
-from motila import motila as mt
+import motila as mt
 # %% VERIFY IMPORT
 mt.hello_world()
 # %% DEFINE MOTILA PARAMETERS
