### 2025-12-02 259d4ba
Remove obsolete .zenodo.json and use CITATION.cff as sole citation source
The .zenodo.json file originated from an earlier DOI workflow and is no longer
needed. Its presence caused conflicting version metadata with CITATION.cff,
as noted by the JOSS reviewer. CITATION.cff is now the single source of truth
for citation information. Zenodo auto-archiving has been disabled accordingly.
