+++ motila/__init__.py
@@ -4,13 +4,15 @@ from .motila import (
     batch_process_stacks,
     batch_collect,
 )
+from .utils import tiff_axes_check_and_correct
 
 __all__ = [
     "hello_world",
     "process_stack",
     "batch_process_stacks",
     "batch_collect",
+    "tiff_axes_check_and_correct"
 ]
 
-# optional: expose the motila module directly
+# expose the motila submodule for backward compatibility
 from . import motila as motila
