### 2025-12-02 d83a253
Update paper.md to clarify MotilA's capabilities for 4D and 5D time-lapse microscopy, enhancing the summary and implementation sections for better understanding.
