### 2025-12-02 bf7ff3c
Add clarification to README on executing example scripts and DATA_Path behavior
This update extends the “Example data set and tutorials” section to explicitly document that example scripts must be executed from within the 'example scripts' directory due to their use of relative paths. A short code snippet illustrates the required workflow, and guidance is provided for users who prefer absolute paths by modifying DATA_Path. This addresses a reviewer comment and improves transparency and usability without altering existing functionality.
