+++ docs/source/overview.rst
@@ -192,6 +192,22 @@ Jupyter notebooks :doc:`demonstrate <tutorials>`:
 Example Python scripts mirror these workflows for use outside of notebooks.
 
 
+Running the example scripts and notebooks
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+The example scripts in ``example scripts`` and ``example notebooks`` expect a relative path 
+layout and therefore must be executed from within that directory. For example:
+
+.. code-block:: bash
+
+   cd example_scripts
+   python single_file_run_paper.py
+
+
+Alternatively, users may modify the ``DATA_Path`` variable inside the script to point 
+to the absolute location of the example project. 
+
+
 .. _license:
 
 License
