### 2025-12-02 5fe5abc
Add initial ROADMAP.md outlining planned improvements and future extensions
