+++ ROADMAP.md
@@ -1,0 +1,20 @@
+# MotilA Roadmap
+
+This roadmap summarizes planned improvements and future extensions derived from user feedback, JOSS reviewer suggestions, and long term development goals. Items listed here are not part of the current release but are planned for future versions.
+
+
+## Support for additional input formats and axis configurations
+* Implement dedicated support for **TYX time-lapse stacks**.
+  The core algorithm is already compatible, but explicit input handling and metadata inference still need to be added.
+
+
+## Data formats and metadata handling
+* Transition all tabular outputs from **.xls** to **.csv** for long term sustainability and interoperability.
+* Consider migrating metadata files (currently .xls) to **.yml**.
+  Migration will preserve backward compatibility and be phased in gradually.
+* Standardize all folder and file names to avoid whitespace and to use underscores for cross-platform stability (e.g., `example dataset` ⟶ `example_dataset`).
+
+
+## Repository organization and auxiliary datasets
+* Provide the **cutout dataset** as a separate Zenodo record linked to the main example dataset.
+  This avoids overwriting example projects and allows independent downloads.
