### 2025-12-02 d90020a
Add initial ROADMAP.md outlining planned improvements and future extensions
