### 2025-12-02 c22ea48
Update README.md to add GitHub last commit and issue badges
