### 2025-12-02 8052a11
Update README.md to add GitHub last commit and issue badges
