### 2025-12-02 6409d74
Update README.md to add GitHub last commit and issue badges
