+++ README.md
@@ -1,8 +1,6 @@
-![GitHub Release](https://img.shields.io/github/v/release/FabrizioMusacchio/motila) [![PyPI version](https://img.shields.io/pypi/v/motila.svg)](https://pypi.org/project/motila/) [![GPLv3 License](https://img.shields.io/badge/License-GPL%20v3-yellow.svg)](https://motila.readthedocs.io/en/latest/overview.html#license) ![Tests](https://github.com/FabrizioMusacchio/motila/actions/workflows/python-tests.yml/badge.svg) [![GitHub last commit](https://img.shields.io/github/last-commit/FabrizioMusacchio/MotilA)](https://github.com/FabrizioMusacchio/MotilA/commits/main/)  [![codecov](https://codecov.io/gh/fabriziomusacchio/motila/branch/main/graph/badge.svg)](https://codecov.io/gh/fabriziomusacchio/motila)  [![GitHub Issues Open](https://img.shields.io/github/issues/FabrizioMusacchio/MotilA)](https://github.com/FabrizioMusacchio/MotilA/issues) [![GitHub Issues Closed](https://img.shields.io/github/issues-closed/FabrizioMusacchio/MotilA)](https://github.com/FabrizioMusacchio/MotilA/issues?q=is%3Aissue%20state%3Aclosed) [![GitHub Issues or Pull Requests](https://img.shields.io/github/issues-pr/FabrizioMusacchio/MotilA)](https://github.com/FabrizioMusacchio/MotilA/pulls)  [![Documentation Status](https://readthedocs.org/projects/motila/badge/?version=latest)](https://motila.readthedocs.io/en/latest/?badge=latest) 
-
-
-
 # MotilA: A pipeline for microglial fine process motility analysis
+
+![GitHub Release](https://img.shields.io/github/v/release/FabrizioMusacchio/motila) [![PyPI version](https://img.shields.io/pypi/v/motila.svg)](https://pypi.org/project/motila/) [![GPLv3 License](https://img.shields.io/badge/License-GPL%20v3-yellow.svg)](https://motila.readthedocs.io/en/latest/overview.html#license) ![Tests](https://github.com/FabrizioMusacchio/motila/actions/workflows/python-tests.yml/badge.svg) [![GitHub last commit](https://img.shields.io/github/last-commit/FabrizioMusacchio/MotilA)](https://github.com/FabrizioMusacchio/MotilA/commits/main/)  [![codecov](https://codecov.io/gh/fabriziomusacchio/motila/branch/main/graph/badge.svg)](https://codecov.io/gh/fabriziomusacchio/motila)  [![GitHub Issues Open](https://img.shields.io/github/issues/FabrizioMusacchio/MotilA)](https://github.com/FabrizioMusacchio/MotilA/issues) [![GitHub Issues Closed](https://img.shields.io/github/issues-closed/FabrizioMusacchio/MotilA)](https://github.com/FabrizioMusacchio/MotilA/issues?q=is%3Aissue%20state%3Aclosed) [![GitHub Issues or Pull Requests](https://img.shields.io/github/issues-pr/FabrizioMusacchio/MotilA)](https://github.com/FabrizioMusacchio/MotilA/pulls)  [![Documentation Status](https://readthedocs.org/projects/motila/badge/?version=latest)](https://motila.readthedocs.io/en/latest/?badge=latest) 
 
 *MotilA* is a Python-based image analysis pipeline designed to quantify microglial fine process motility from 4D and 5D time-lapse image stacks acquired through multi-photon in vivo imaging. While developed for microglial analysis, *MotilA* can be applied to other cell types and imaging studies as well. The pipeline supports both single-file and batch processing, making it adaptable for various experimental designs and high-throughput analyses. 
 
