### 2025-12-03 557f82b
Update README.md and index.rst to enhance badge visibility and add new metrics
