### 2025-12-03 0f6468f
Update README.md and index.rst to enhance badge visibility and add new metrics
