### 2025-12-03 77ba1ef
Update README.md and index.rst to enhance badge visibility and add new metrics
