### 2025-12-03 433fb14
Update README.md and index.rst to enhance badge visibility and add new metrics
