### 2025-12-03 c857786
Update README.md and index.rst to enhance badge visibility and add new metrics
