### 2025-12-03 cce19d6
Update README.md and index.rst to enhance badge visibility and add new metrics
