### 2025-12-03 eeafb1e
Update ROADMAP.md: added  _processing_parameters.xlsx to the yaml transition target list
