+++ motila/__init__.py
@@ -4,14 +4,28 @@ from .motila import (
     batch_process_stacks,
     batch_collect,
 )
-from .utils import tiff_axes_check_and_correct
+from .utils import (
+    tiff_axes_check_and_correct,
+    check_folder_exist_create,
+    filterfolder_by_string,
+    filterfiles_by_string,
+    logger_object,
+    print_ram_usage_in_loop,
+    print_ram_usage
+)
 
 __all__ = [
     "hello_world",
     "process_stack",
     "batch_process_stacks",
     "batch_collect",
-    "tiff_axes_check_and_correct"
+    "tiff_axes_check_and_correct",
+    "check_folder_exist_create",
+    "filterfolder_by_string",
+    "filterfiles_by_string",
+    "logger_object",
+    "print_ram_usage_in_loop",
+    "print_ram_usage",
 ]
 
 # expose the motila submodule for backward compatibility
