### 2025-12-03 5cc6de2
Update example dataset citations and README documentation
- Changed citations in README.md, tutorials.rst, and example notebooks to reflect the correct (zenodo!) authorship (Musacchio et al., 2025).
- Added information about a smaller "cutout" version of the example dataset in README.md and tutorials.rst.
- Updated the example project README to ensure consistency in dataset citation.
