+++ example project/README.md
@@ -2,6 +2,6 @@
 
 **Before you proceed, please make sure that you have downloaded the example data from Zenodo:** 
 
-* [Gockel &  Nieves-Rivera (2025), doi: 10.5281/zenodo.15061566](https://zenodo.org/records/15061566)
+* [Musacchio et al. (2025), doi: 10.5281/zenodo.15061566](https://zenodo.org/records/15061566)
 
 Place the downloaded and extracted data set into this folder (`example project`).
