### 2025-12-03 4336758
Update roadmap to indicate cutout dataset provision addressed in release v1.1.0
