### 2025-12-03 bfa5f3e
Update roadmap to indicate cutout dataset provision addressed in release v1.1.0
