### 2025-12-03 b5ec239
Bump version to 1.1.0 in pyproject.toml
