### 2025-12-04 5a3fe56
Add Zenodo Archive and bioRxiv Preprint badges to documentation
