### 2025-12-05 27f5e9f
Update README and documentation to include total downloads badge for PyPI
