### 2025-12-05 f79474a
Update README and documentation to include total downloads badge for PyPI
