### 2025-12-05 4c916a7
Update README and documentation to include total downloads badge for PyPI
