### 2025-12-09 9306cea
Update dataset citation in README and notebooks
- Revised citation for the full example dataset in README.md, tutorials.rst, single_file_run.ipynb, and batch_run.ipynb to include Gockel & Nieves-Rivera as co-authors.
- Ensured consistency across documentation regarding dataset references.
