### 2025-12-13 6cccd23
Paper: fixed formatting ("{}") and updated Zarr citation to correct version 3.1.5 in bibliography
