+++ paper/paper.bib
@@ -36,7 +36,7 @@
 	journal = {Nature Neuroscience},
 	number = {4},
 	pages = {411--413},
-	title = {Microglial Cx3cr1 knockout prevents neuron loss in a mouse model of Alzheimer's disease},
+	title = {Microglial Cx3cr1 knockout prevents neuron loss in a mouse model of {Alzheimer's} disease},
 	url = {https://doi.org/10.1038/nn.2511},
 	volume = {13},
 	year = {2010},
@@ -194,8 +194,8 @@ doi = {10.1364/OL.33.000156},
 
 % scipy and numpy package:
 @article{Virtanen:2020,
-  title     = {SciPy 1.0: Fundamental Algorithms for Scientific Computing in Python},
-  author    = {Pauli Virtanen and Ralf Gommers and Travis E. Oliphant and Matt Haberland and Tyler Reddy and David Cournapeau and Evgeni Burovski and Pearu Peterson and Warren Weckesser and Jonathan Bright and St{\'e}fan J. van der Walt and Matthew Brett and Joshua Wilson and K. Jarrod Millman and Nikolay Mayorov and Andrew R. J. Nelson and Eric Jones and Robert Kern and Eric Larson and C J Carey and {\.I}lhan Polat and Yu Feng and Eric W. Moore and Jake VanderPlas and Denis Laxalde and Josef Perktold and Robert Cimrman and Ian Henriksen and E. A. Quintero and Charles R. Harris and Anne M. Archibald and Ant{\^o}nio H. Ribeiro and Fabian Pedregosa and Paul van Mulbregt and SciPy 1.0 Contributors},
+  title     = {{SciPy} 1.0: Fundamental Algorithms for Scientific Computing in {Python}},
+  author    = {Pauli Virtanen and Ralf Gommers and Travis E. Oliphant and Matt Haberland and Tyler Reddy and David Cournapeau and Evgeni Burovski and Pearu Peterson and Warren Weckesser and Jonathan Bright and St{\'e}fan J. {van der Walt} and Matthew Brett and Joshua Wilson and K. Jarrod Millman and Nikolay Mayorov and Andrew R. J. Nelson and Eric Jones and Robert Kern and Eric Larson and C J Carey and {\.I}lhan Polat and Yu Feng and Eric W. Moore and Jake VanderPlas and Denis Laxalde and Josef Perktold and Robert Cimrman and Ian Henriksen and E. A. Quintero and Charles R. Harris and Anne M. Archibald and Ant{\^o}nio H. Ribeiro and Fabian Pedregosa and Paul van Mulbregt and {SciPy 1.0 Contributors}},
   journal   = {Nature Methods},
   volume    = {17},
   pages     = {261--272},
@@ -206,7 +206,7 @@ doi = {10.1364/OL.33.000156},
 
 @article{Harris:2020,
   title     = {Array programming with {NumPy}},
-  author    = {Charles R. Harris and K. Jarrod Millman and St{\'e}fan J. van der Walt and Ralf Gommers and Pauli Virtanen and David Cournapeau and Eric Wieser and Julian Taylor and Sebastian Berg and Nathaniel J. Smith and Robert Kern and Matti Picus and Stephan Hoyer and Marten H. van Kerkwijk and Matthew Brett and Alec Haldane and Jaime Fern{\'a}ndez del R{\'i}o and Mark Wiebe and Pearu Peterson and Pierre G{\'e}rard-Marchant and Kevin Sheppard and Tyler Reddy and Warren Weckesser and Hameer Abbasi and Christoph Gohlke and Travis E. Oliphant},
[…]
