+++ paper/paper.md
@@ -42,7 +42,7 @@ bibliography: paper.bib
 
 
 ## Summary
-*MotilA* is an open-source Python pipeline for quantifying microglial fine-process motility in 4D (TZYX) or 5D (TZCYX) time-lapse fluorescence microscopy data, supporting both single-channel and two-channel acquisition. It was developed for high-resolution *in vivo* multiphoton imaging and supports both single-stack and cohort-scale batch analyses. The workflow performs sub-volume extraction, optional registration and spectral unmixing, a maximum-intensity projection along the Z-axis, segmentation, and pixel-wise change detection to compute the turnover rate (TOR). *MotilA* specifically targets pixel-level process motility rather than object tracking or full morphometry.  The code is platform independent, documented with tutorials and example datasets, and released under GPL-3.0.
+*MotilA* is an open-source Python pipeline for quantifying microglial fine-process motility in 4D (TZYX) or 5D (TZCYX) time-lapse fluorescence microscopy data, supporting both single-channel and two-channel acquisition. It was developed for high-resolution *in vivo* multiphoton imaging and supports both single-stack and cohort-scale batch analyses. The workflow performs sub-volume extraction, optional registration and spectral unmixing, a maximum-intensity projection along the Z-axis, segmentation, and pixel-wise change detection to compute the turnover rate (TOR). *MotilA* specifically targets pixel-level process motility rather than object tracking or full morphometry. The code is platform independent, documented with tutorials and example datasets, and released under GPL-3.0.
 
 
 ## Statement of need
@@ -92,18 +92,12 @@ The pipeline exposes options for 3D or 2D registration, contrast-limited adaptiv
 *MotilA* has been applied to quantify microglial process dynamics in several *in vivo* imaging studies and preprints [@FFuhrmann:2024; @Crux:2024; @Gockel:2025]. Typical use cases include baseline surveillance behavior, responses to neuroinflammation or genetic perturbations, and deep three-photon imaging where manual analysis is impractical. The binarize-and-compare principle can in principle be adapted to other structures such as dendrites or axons when segmentation across time is robust.
 
 ## Limitations
-Using 2D projections simplifies processing but sacrifices axial specificity and can merge overlapping structures. Segmentation quality determines accuracy and can be affected by vessels, low signal-to-noise ratios, or strong intensity drift. The current spectral unmixing is a simple subtraction; advanced approaches may be needed for some fluorophores. *MotilA* targets pixel-level process motility rather than object-level tracking or full morphometry.
+*MotilA* quantifies microglial process motility using 2D maximum-intensity projections rather than fully volumetric analysis. This choice reflects practical constraints of two-photon microglial imaging: axial resolution degrades with depth, producing elongated point-spread functions and reduced contrast along Z, which makes reliable voxel-wise 3D segmentation of thin processes difficult. Z-projection increases effective signal per pixel and follows established practice in earlier microglial motility studies (see, e.g., @Fuhrmann:2010; @Nebeling:2023), but necessarily sacrifices axial specificity and may merge structures that overlap along Z. Users are therefore advised to select sub-volumes with minimal axial overlap.
 
[…]
