### 2025-12-13 f712e50
Fix citation year for Helmchen article from 2001 to 2005 in bibliography
