### 2025-12-19 #9267 @csoneson
☝🏻 @FabrizioMusacchio I think this comes from the `Fuhrmann:2010` bib entry - the indicated ISBN is actually an ISSN
