### 2025-12-20 e729dcc
Clean up bibliography entries by removing unnecessary fields and correcting author formatting
