+++ paper/paper.bib
@@ -31,8 +31,6 @@
   author = {Martin Fuhrmann and Tobias Bittner and Christian K. E. Jung and Steffen Burgold and Richard M. Page and Gerda Mitteregger and Christian Haass and Frank M. LaFerla and Hans Kretzschmar and Jochen Herms},
 	date = {2010/04/01},
 	doi = {10.1038/nn.2511},
-	id = {Fuhrmann2010},
-	isbn = {1546-1726},
 	journal = {Nature Neuroscience},
 	number = {4},
 	pages = {411--413},
@@ -56,9 +54,7 @@
   citation = {eLife 2023;12:e83176},
   doi = {10.7554/eLife.83176},
   url = {https://doi.org/10.7554/eLife.83176},
-  keywords = {microglia, hippocampus, two-photon, dendritic spines, chemogenetics},
   journal = {eLife},
-  issn = {2050-084X},
   publisher = {eLife Sciences Publications, Ltd},
 }
 
@@ -166,7 +162,7 @@
   address   = {New York, NY, USA},
 }
 
-@ARTICLE{Anuta:1970,
+@article{Anuta:1970,
   author={Anuta, Paul E.},
   journal={IEEE Transactions on Geoscience Electronics}, 
   title={Spatial Registration of Multispectral and Multitemporal Digital Imagery Using Fast Fourier Transform Techniques}, 
@@ -175,7 +171,8 @@
   number={4},
   pages={353-368},
   keywords={Digital images;Fast Fourier transforms;Satellites;Optical filters;Image analysis;Photography;Space technology;Multispectral imaging;Cameras;Optical sensors},
-  doi={10.1109/TGE.1970.271435}}
+  doi={10.1109/TGE.1970.271435}
+}
 
 @article{GuizarSicairos:2008,
 author = {Manuel Guizar-Sicairos and Samuel T. Thurman and James R. Fienup},
@@ -270,7 +267,7 @@ doi = {10.1364/OL.33.000156},
 % thresholding methods:
 @article{Ridler:1978,
   title     = {Picture thresholding using an iterative selection method},
-  author    = {T. W. Ridler and S. Calvard and Full name},
+  author    = {T. W. Ridler and S. Calvard},
   journal   = {IEEE Transactions on Systems, Man, and Cybernetics},
   volume    = {8},
   number    = {8},
@@ -363,7 +360,7 @@ doi = {10.1364/OL.33.000156},
   institution = {Lawrence Berkeley National Laboratory, University of California},
   year        = {2005},
   number      = {LBNL-56864},
-  url         = {http://repositories.cdlib.org/lbnl/LBNL-56864},
+  url         = {https://doi.org/10.1117/12.596105},
   type        = {Technical Report},
   doi         = {10.1117/12.596105},
 }
