### 2025-12-22 4014107
Update citation for Fuhrmann et al. article: change year to 2025, add volume, number, and pages, and update journal and publisher details.
