+++ paper/paper.bib
@@ -141,14 +141,15 @@
 @article {FFuhrmann:2024,
   author = {Falko Fuhrmann and Felix C. Nebeling and Fabrizio Musacchio and Manuel Mittag and Stefanie Poll and Monika M{\"u}ller and Eleonora Ambrad Giovannetti and Michael Maibach and Barbara Schaffran and Emily Burnside and Ivy Chi Wai Chan and Alex Simon Lagurin and Nicole Reichenbach and Sanjeev Kaushalya and Hans Fried and Stefan Linden and Gabor C. Petzold and Gaia Tavosanis and Frank Bradke and Martin Fuhrmann},
 	title = {Three-photon in vivo imaging of neurons and glia in the medial prefrontal cortex with sub-cellular resolution},
-	elocation-id = {2024.08.28.610026},
-	year = {2024},
-	doi = {10.1101/2024.08.28.610026},
-	publisher = {Cold Spring Harbor Laboratory},
-	URL = {https://www.biorxiv.org/content/early/2024/08/28/2024.08.28.610026},
-	eprint = {https://www.biorxiv.org/content/early/2024/08/28/2024.08.28.610026.full.pdf},
-	journal = {bioRxiv},
-  note = {Fuhrmann, Nebelung, and Musacchio contributed equally to this work.},
+	year = {2025},
+  volume = {8},
+  number = {1},
+  pages={795},
+	doi = {10.1038/s42003-025-08079-8},
+	URL = {https://doi.org/10.1038/s42003-025-08079-8},
+	journal = {Communications Biology},
+  publisher={Nature Publishing Group UK London},
+  note = {Fuhrmann, Nebeling, and Musacchio contributed equally to this work.},
   shortauthor = {Fuhrmann, Nebeling, Musacchio et al.}
 }
 
