### 2025-12-23 cf74a57
Fix formatting in CONTRIBUTING.md by correcting code block syntax
