### 2025-12-23 29b7466
Update section header formatting in overview.rst for consistency
