### 2025-12-23 d122f12
fix: RTD build API function auto-build was broken due to server-site sphinx updates/theme outdated: fixed
