+++ docs/requirements.txt
@@ -1,3 +1,4 @@
-sphinx
-sphinx_rtd_theme
-sphinx-autodoc-typehints
+# docs/requirements.txt
+sphinx>=8,<10
+sphinx-rtd-theme>=3.0.0
+sphinx-autodoc-typehints>=2.0.0
