### 2025-12-23 f5fa70c
fix: correct Python installation method in Read the Docs configuration
