+++ .readthedocs.yaml
@@ -20,5 +20,6 @@ sphinx:
 # See https://docs.readthedocs.io/en/stable/guides/reproducible-builds.html
 python:
   install:
+    - method: pip
+      path: .
     - requirements: docs/requirements.txt
-        
