### 2025-12-24 bd664e2
fix: update citation information to reflect peer-reviewed publication in JOSS
