### 2025-12-24 6b2de1a
chore: remove draft PDF workflow from GitHub Actions
