+++ .github/workflows/draft-pdf.yml
@@ -1,52 +1,0 @@
-name: Draft PDF
-on: [push]
-
-permissions:
-  contents: write  # Grant write access to repository contents
-
-# Verify Repository Settings
-# Ensure that the repository allows workflows to use the GITHUB_TOKEN with write permissions:
-
-# 1. Go to your repository on GitHub.
-# 2. Navigate to Settings > Actions > General.
-# 3. Under Workflow permissions, select Read and write permissions.
-# 4. Save the changes.
-
-jobs:
-  paper:
-    runs-on: ubuntu-latest
-    name: Paper Draft
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v4
-
-      - name: Check for paper folder and paper.md
-        id: check-paper
-        run: |
-          success=true
-          if [ -d "paper" ] && [ -f "paper/paper.md" ]; then
-            echo "Paper folder and paper.md exist."
-          else
-            echo "Paper folder or paper.md is missing."
-            success=false
-          fi
-          echo "success=$success" >> $GITHUB_ENV  # Export the variable to the environment
-
-      - name: Build draft PDF
-        if: env.success == 'true'  # Check the success variable
-        uses: openjournals/openjournals-draft-action@master
-        with:
-          journal: joss
-          paper-path: paper/paper.md
-
-      - name: Commit and push PDF to repository
-        if: env.success == 'true'  # Check the success variable
-        run: |
-          git config --global user.name "github-actions[bot]"
-          git config --global user.email "github-actions[bot]@users.noreply.github.com"
-          #TIMESTAMP=$(date +"%y%m%d_%H%M") # use is by adding ${TIMESTAMP} to the filename
-          mv paper/paper.pdf paper/Musacchio_MotilA_paper_draft.pdf
-          git add paper/Musacchio_MotilA_paper_draft.pdf
-          git commit -m "Add draft PDF with timestamp: Musacchio_MotilA_paper_draft.pdf [skip ci]"
-          git push
-
