### 2025-12-24 465cbce
fix: resolve omio version dynamically in Sphinx configuration
