+++ docs/source/conf.py
@@ -11,10 +11,32 @@ import sys
 from datetime import datetime
 sys.path.insert(0, os.path.abspath("../.."))  # ../.. = where motila/ is located
 
+from importlib.metadata import version as pkg_version, PackageNotFoundError, packages_distributions
+
+def _resolve_omio_version() -> str:
+    # primary: known PyPI distribution name
+    try:
+        return pkg_version("omio-microscopy")
+    except PackageNotFoundError:
+        pass
+
+    # fallback: map import package -> installed distribution(s)
+    try:
+        dist_names = packages_distributions().get("omio", [])
+        for dist in dist_names:
+            try:
+                return pkg_version(dist)
+            except PackageNotFoundError:
+                continue
+    except Exception:
+        pass
+
+    return "0.0.0+unknown"
+
 project = 'MotilA'
 copyright = '2025, Fabrizio Musacchio'
 author = 'Fabrizio Musacchio'
-release = 'v1.0.7'
+release = _resolve_omio_version()
 copyright = f"{datetime.now().year}, {author}"
 
 # -- General configuration ---------------------------------------------------
