### 2025-12-24 4b29098
docs: update README and index.rst to reflect changes in badge links and remove outdated bioRxiv reference
