### 2025-12-24 c7b8fee
docs: update commented bioRxiv badge link in README
