### 2025-12-24 e06c7be
docs: add animation figure illustrating MotilA's segmentation and microglial turnover rate assessment
