### 2025-12-28 1ec8fd5
docs: remove outdated comments from index.rst
