+++ docs/source/index.rst
@@ -1,8 +1,3 @@
-.. MotilA documentation master file, created by
-   sphinx-quickstart on Fri Nov 21 14:13:59 2025.
-   You can adapt this file completely to your liking, but it should at least
-   contain the root `toctree` directive.
-
 MotilA documentation
 ====================
 
